streamladder/requirements.txt
playwright
python-dotenv
requests
loguru
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
boto3
